Refactor : move LoginHelpers to Util package
